+++ examples/sandbox/variables.tf
@@ -1,0 +1,25 @@
+# Copyright 2026 Boozt Fashion AB
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+variable "project_id" {
+  description = "GCP project ID."
+  type        = string
+}
+
+variable "region" {
+  description = "GCP region."
+  type        = string
+  default     = "europe-west1"
+}
+
+variable "organization_id" {
+  description = "GCP organization ID for Cloud Asset Inventory integration."
+  type        = string
+  default     = ""
+}
+